requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.1.8
Flask==3.1.2
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
typing-inspection==0.4.1
typing_extensions==4.15.0
Werkzeug==3.1.3
zipp==3.23.0
